エンドポイントにv1を追加 & 構成を簡略化
