--- src/app/api/_middlewares/core.ts
@@ -1,5 +1,0 @@
-import { createFactory } from 'hono/factory'
-
-import type { Env } from '../types'
-
-export const factory = createFactory<Env>()
