+++ src/app/api/_router/v1/utils.ts
@@ -1,0 +1,11 @@
+import { createFactory } from 'hono/factory'
+
+import type { User } from '@prisma/client'
+
+type Env = {
+  Variables: {
+    user: User
+  }
+}
+
+export const factory = createFactory<Env>()
