--- src/app/api/core.ts
@@ -1,22 +1,0 @@
-import { Hono } from 'hono'
-import { logger } from 'hono/logger'
-import { prettyJSON } from 'hono/pretty-json'
-import { handle } from 'hono/vercel'
-
-import { embed } from './_router/embed'
-import { ogp } from './_router/ogp'
-import { skill } from './_router/skills'
-import { tag } from './_router/tags'
-import { work } from './_router/works'
-
-const app = new Hono()
-  .basePath('/api')
-  .route('/embeds', embed)
-  .route('/ogp', ogp)
-  .route('/skills', skill)
-  .route('/tags', tag)
-  .route('/works', work)
-  .use('*', prettyJSON(), logger())
-
-export const handler = handle(app)
-export type AppType = typeof app
