--- src/app/api/types.ts
@@ -1,7 +1,0 @@
-import type { User } from '@prisma/client'
-
-export type Env = {
-  Variables: {
-    user: User
-  }
-}
